findbugs fixes
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1735319 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/poifs/storage/BATBlock.java
@@ -242,12 +242,13 @@ public final class BATBlock extends BigBlock {
      */
     public static long calculateMaximumSize(final POIFSBigBlockSize bigBlockSize,
           final int numBATs) {
-       long size = 1; // Header isn't FAT addressed
+       // Header isn't FAT addressed
+       long size = 1;
        
        // The header has up to 109 BATs, and extra ones are referenced
        //  from XBATs
        // However, all BATs can contain 128/1024 blocks
-       size += (numBATs * bigBlockSize.getBATEntriesPerBlock());
+       size += (((long)numBATs) * bigBlockSize.getBATEntriesPerBlock());
        
        // So far we've been in sector counts, turn into bytes
        return size * bigBlockSize.getBigBlockSize();
